Fixes to make the new functionality we've added opt-in (#6)
* Fixes to make the new functionality we've added opt-in

Namely, adds the following two new DSL keywords:

- ember_pouch_mode
- upsert_mode

These are pretty self-explanatory, and do what one would expect. They are _disabled_ by default, meaning we retain the `couch_tap` default behavior.

Since this makes the tests also work again, this PR closes #5.

* Fixes to make Travis happy

This brings in an update to CouchRest to 2.x, which should be fine. I've tested this commit both locally and with the unit tests, seems to work fine.