Use standard JSON library instead of yajl
yajl-ruby has the downside that is relies on a C extension which works poorly (well, not at all...) with JRuby. We are targetting JRuby in our main project using this gem, so this change is mandatory for us to be able to use it.

I've tested this locally and it seems to stream changes fine. However, it's only been tested with very small amounts of data as of yet.